Removing createAndRunQSUB from Simulation.py
